refactor(number-theory): isolate prime contracts
